--- docker-compose.yml
@@ -1,10 +1,0 @@
-version: '3'
-
-services:
-  pdf-extractor:
-    build: .
-    volomes:
-      - ${PDF_SOURCE_DIR}:/app/documents
-      - ${PDF_OUTPUT_DIR:}:/app/output
-    environment:
-      - PYTHONUNBUFFERED=1
