+++ pdf_processor_settings.json
@@ -1,8 +1,8 @@
 {
-  "input_path": "",
+  "input_path": "/Users/dingu/Documents/JS_FM",
   "image_output_dir": "/Users/dingu/Note/obsidian/03_assets/本表紙",
   "symlink_output_dir": "/Users/dingu/Note/obsidian/書籍",
-  "subdir_name": "book_covers",
+  "subdir_name": "JS+α",
   "use_table": true,
   "show_title": false
 }
