--- requirements.txt
@@ -1,4 +1,0 @@
-PyPDF2==2.12.1
-pdf2image==1.16.3
-pillow==9.4.0
-pyinstaller==5.9.0
